+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ordemservico/OrdemServicoDaoImpl.java
@@ -71,7 +71,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
     @Override
     public List<OrdemServicoListagem> getOrdemServicoListagem(@NotNull final Long codUnidade,
                                                               @Nullable final Long codTipoVeiculo,
-                                                              @Nullable final String placa,
+                                                              @Nullable final Long codVeiculo,
                                                               @Nullable final StatusOrdemServico statusOrdemServico,
                                                               final int limit,
                                                               final int offset) throws Throwable {
@@ -83,7 +83,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
             stmt = conn.prepareStatement("SELECT * FROM FUNC_CHECKLIST_OS_GET_OS_LISTAGEM(?, ?, ?, ?, ?, ?)");
             stmt.setLong(1, codUnidade);
             bindValueOrNull(stmt, 2, codTipoVeiculo, SqlType.BIGINT);
-            bindValueOrNull(stmt, 3, placa, SqlType.TEXT);
+            bindValueOrNull(stmt, 3, codVeiculo, SqlType.BIGINT);
             if (statusOrdemServico != null) {
                 stmt.setString(4, statusOrdemServico.asString());
             } else {
@@ -107,7 +107,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
     public List<QtdItensPlacaListagem> getQtdItensPlacaListagem(
             @NotNull final Long codUnidade,
             @Nullable final Long codTipoVeiculo,
-            @Nullable final String placaVeiculo,
+            @Nullable final Long codVeiculo,
             @Nullable final StatusItemOrdemServico statusItens,
             final int limit,
             final int offset) throws Throwable {
@@ -120,7 +120,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
                     "SELECT * FROM FUNC_CHECKLIST_OS_GET_QTD_ITENS_PLACA_LISTAGEM(?, ?, ?, ?, ?, ?)");
             stmt.setLong(1, codUnidade);
             bindValueOrNull(stmt, 2, codTipoVeiculo, SqlType.BIGINT);
-            bindValueOrNull(stmt, 3, placaVeiculo, SqlType.TEXT);
+            bindValueOrNull(stmt, 3, codVeiculo, SqlType.BIGINT);
             if (statusItens != null) {
                 stmt.setString(4, statusItens.asString());
             } else {
@@ -167,7 +167,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
     @NotNull
     @Override
     public HolderResolucaoItensOrdemServico getHolderResolucaoItensOrdemServico(
-            @NotNull final String placaVeiculo,
+            @NotNull final Long codVeiculo,
             @Nullable final PrioridadeAlternativa prioridade,
             @Nullable final StatusItemOrdemServico statusItens,
             final int limit,
@@ -182,7 +182,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
             stmt.setNull(1, SqlType.BIGINT.asIntTypeJava());
             // Código da Ordem de Serviço.
             stmt.setNull(2, SqlType.BIGINT.asIntTypeJava());
-            stmt.setString(3, placaVeiculo);
+            stmt.setLong(3, codVeiculo);
             if (prioridade != null) {
                 stmt.setString(4, prioridade.asString());
             } else {
@@ -212,7 +212,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
     public HolderResolucaoItensOrdemServico getHolderResolucaoMultiplosItens(
             @Nullable final Long codUnidade,
             @Nullable final Long codOrdemServico,
-            @Nullable final String placaVeiculo,
+            @Nullable final Long codVeiculo,
             @Nullable final StatusItemOrdemServico statusItens) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
@@ -222,7 +222,7 @@ public final class OrdemServicoDaoImpl extends DatabaseConnection implements Ord
             stmt = conn.prepareStatement("SELECT * FROM FUNC_CHECKLIST_OS_GET_ITENS_RESOLUCAO(?, ?, ?, ?, ?, ?, ?, ?)");
             bindValueOrNull(stmt, 1, codUnidade, SqlType.BIGINT);
             bindValueOrNull(stmt, 2, codOrdemServico, SqlType.BIGINT);
-            bindValueOrNull(stmt, 3, placaVeiculo, SqlType.TEXT);
+            bindValueOrNull(stmt, 3, codVeiculo, SqlType.BIGINT);
             stmt.setNull(4, SqlType.TEXT.asIntTypeJava());
             bindValueOrNull(stmt, 5, statusItens != null ? statusItens.asString() : null, SqlType.VARCHAR);
             stmt.setObject(6, OffsetDateTime.now(Clock.systemUTC()));